04_파이썬웹개발/디자인자동화/requirements.txt
faker==19.13.0
python-dotenv==1.0.0
requests==2.31.0
figma-api==0.9.1